c
c

+++ MyApp.Domain/Entities/Product.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace MyApp.Domain.Entities;
 
 public class Product : BaseEntity
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public Int64 Price { get; set; }
 }
